Add signup styles
