+++ src/components/Auth/SignIn/SignIn.test.js
@@ -2,87 +2,75 @@ import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 
 import { fillInput, fireEvent, render, wait } from 'testUtils';
-import { mockSignUpSuccess, mockSignUpFailure } from 'testUtils/mocks/auth';
+import { mockSignInSuccess, mockSignInFailure } from 'testUtils/mocks/auth';
 import fakeAuthService from 'api/AuthService';
-import SignUp from '.';
+import SignIn from '.';
 
 jest.mock('api/AuthService');
 
 const fakeCredentials = {
   email: 'user@example.com',
-  firstName: 'User',
-  lastName: 'Example',
   password: 'password',
 };
 
-describe('SignUp', () => {
+describe('SignIn', () => {
   it('should submit correctly', async () => {
-    mockSignUpSuccess();
+    mockSignInSuccess();
 
-    const { getByTestId } = render(<SignUp />);
-    const email = getByTestId('email-input');
-    const firstName = getByTestId('firstName-input');
-    const lastName = getByTestId('lastName-input');
-    const password = getByTestId('password-input');
+    const { getByTestId } = render(<SignIn />);
     const submitButton = getByTestId('submit-button');
+    const email = getByTestId('email-input');
+    const password = getByTestId('password-input');
 
     await wait(() => {
       fillInput(email, fakeCredentials.email);
-      fillInput(firstName, fakeCredentials.firstName);
-      fillInput(lastName, fakeCredentials.lastName);
       fillInput(password, fakeCredentials.password);
     });
 
     expect(submitButton).toBeEnabled();
     expect(email.value).toBe(fakeCredentials.email);
-    expect(firstName.value).toBe(fakeCredentials.firstName);
-    expect(lastName.value).toBe(fakeCredentials.lastName);
     expect(password.value).toBe(fakeCredentials.password);
 
     await wait(() => {
       fireEvent.click(submitButton);
     });
 
-    expect(fakeAuthService.signUp).toHaveBeenCalledTimes(1);
-    expect(fakeAuthService.signUp).toHaveBeenCalledWith(fakeCredentials);
+    expect(fakeAuthService.signIn).toHaveBeenCalledTimes(1);
+    expect(fakeAuthService.signIn).toHaveBeenCalledWith(fakeCredentials);
   });
 
   it('should show error on response failure', async () => {
-    mockSignUpFailure();
+    mockSignInFailure();
 
-    const { getByTestId, queryByText } = render(<SignUp />);
-    const email = getByTestId('email-input');
-    const firstName = getByTestId('firstName-input');
-    const lastName = getByTestId('lastName-input');
-    const password = getByTestId('password-input');
+    const { getByTestId } = render(<SignIn />);
     const submitButton = getByTestId('submit-button');
+    const email = getByTestId('email-input');
+    const password = getByTestId('password-input');
 
     await wait(() => {
       fillInput(email, fakeCredentials.email);
-      fillInput(firstName, fakeCredentials.firstName);
-      fillInput(lastName, fakeCredentials.lastName);
       fillInput(password, fakeCredentials.password);
     });
 
     expect(submitButton).toBeEnabled();
     expect(email.value).toBe(fakeCredentials.email);
-    expect(firstName.value).toBe(fakeCredentials.firstName);
-    expect(lastName.value).toBe(fakeCredentials.lastName);
     expect(password.value).toBe(fakeCredentials.password);
 
     await wait(() => {
       fireEvent.click(submitButton);
     });
 
-    expect(fakeAuthService.signUp).toHaveBeenCalledTimes(1);
-    expect(fakeAuthService.signUp).toHaveBeenCalledWith(fakeCredentials);
-    expect(queryByText('There was an error.')).toBeInTheDocument();
+    expect(fakeAuthService.signIn).toHaveBeenCalledTimes(1);
+    expect(fakeAuthService.signIn).toHaveBeenCalledWith(fakeCredentials);
+
+    const errorMessage = getByTestId('error-message');
+    expect(errorMessage.innerHTML).toBe('An error occurred');
   });
 
   it('should disable the submit button for invalid values', async () => {
-    const { getByTestId } = render(<SignUp />);
-    const email = getByTestId('email-input');
+    const { getByTestId } = render(<SignIn />);
     const submitButton = getByTestId('submit-button');
+    const email = getByTestId('email-input');
 
     await wait(() => {
       fillInput(email, 'invalid');
