+++ src/languageProvider/messages/en-US.js
@@ -8,5 +8,7 @@ export default {
     signUp: 'Sign up',
     dontHaveAccountQuestion: "Don't have an account?",
     createAccount: 'Create one here',
+    alreadyHaveAnAccount: 'Already have an account?',
+    errorOccurred: 'An error occurred',
   },
 };
