+++ src/languageProvider/messages/es-ES.js
@@ -8,5 +8,7 @@ export default {
     signUp: 'Crear Cuenta',
     dontHaveAccountQuestion: '¿No tienes cuenta?',
     createAccount: 'Crea una aquí',
+    alreadyHaveAnAccount: 'Ya tienes una cuenta?',
+    errorOccurred: 'Ha ocurrido un error',
   },
 };
